Extract the legend from the first figure, remove the legends from all plots, and put the legend at the end.

Extract the legend from the first figure, remove the legends from all plots, and put the legend at the end.

Usage

plot_grid_one_legend(...)

Arguments

...

One or more ggplot2 objects

Value

A list of the ggplot2 objects (suitable for the plotlist argument of cowplot::plot_grid)

See also